requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
piexif>=1.1.3
croniter>=1.0.0
mutagen>=1.45.0
filelock>=3.0.0
